% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3control_operations.R
\name{get_public_access_block}
\alias{get_public_access_block}
\title{Retrieves the Public Access Block configuration for an Amazon Web Services account}
\usage{
get_public_access_block(AccountId)
}
\arguments{
\item{AccountId}{[required] The Account ID for the Amazon Web Services account whose Public Access Block configuration you want to retrieve.}
}
\description{
Retrieves the Public Access Block configuration for an Amazon Web Services account.
}
\section{Accepted Parameters}{
\preformatted{get_public_access_block(
  AccountId = "string"
)
}
}

