% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3control_package.R
\docType{package}
\name{paws.s3control-package}
\alias{paws.s3control}
\alias{paws.s3control-package}
\title{paws.s3control: AWS S3 Control}
\description{
AWS S3 Control provides access to Amazon S3 control plane
operations.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
